auth: Moved `oauth` packages under `auth`.
The new `auth` package contains all functionality related to
credentials and authentication. OAuth is just a special case of this;
therefore, this package is now a subpackage of `auth`.
