+++ src/integrationTest/scala/com/github/sync/cli/oauth/OAuthLoginCommandSpec.scala
@@ -19,8 +19,8 @@ package com.github.sync.cli.oauth
 import com.github.cloudfiles.core.http.auth.{OAuthConfig, OAuthTokenData}
 import com.github.cloudfiles.core.http.{HttpRequestSender, Secret}
 import com.github.scli.ConsoleReader
+import com.github.sync.auth.oauth.{IDPConfig, OAuthStorageService, OAuthTokenRetrieverService, OAuthTokenRetrieverServiceImpl, SyncOAuthStorageConfig}
 import com.github.sync.cli.oauth.OAuthParameterManager.LoginCommandConfig
-import com.github.sync.oauth.*
 import com.github.sync.{AsyncTestHelper, WireMockSupport}
 import com.github.tomakehurst.wiremock.client.WireMock.*
 import org.apache.pekko.Done
